Eqtide and Thermint now report the same value for tidal power, at least for the first step. The old thermint code used dRotRate to calculate Im(k_2) instead of dMeanMotion. I also changed all the PropertiesX to PropsAuxX, which will cause some compilation errors. This change affects vplanet.h, control.c, module.c and all the module files. It was necessary so there is no confusion regarding a potential difference between "Properties" and "AuxProps".
I am only able to get the log files to reproduce the same TidalPower values as I can't run an integration yet. Peter gave me an updated version of BodyCopyThermint (which is in this version), but many of the body elements don't seem to exist in vplanet.h. So this version is currently broken, and needs to be merged with Peter's.