Several new features added:
New module FLARE has been added. At this time, it is primarily just a placeholder for future more sophisticated treatments. It should be ignored.

I also moved some items around in STELLAR to accomodate new FLARE features. Most notable is that I move PropertiesStellar to a new function called fdLXUVStellar. The original function was not necessary for STELLAR to work, so I think a separate function that is called only during Write is sufficient.

Finally I changed thermint so that the inner core can not grow larger than the core itself (I think).