Fixed bug in which SurfEnFluxTot was converted to the wrong units (it was still in cgs).
Fixed bug in which FinalizeOutputThermint was setting the SurfEnFLux function to the RADHEAT function.

Added RadPowerCrust to RADHEAT which always return 7 TW. The current Earth produces 7 TW of of radiogenic power in the crust.
I will wave my hands and say that since Earth's crust has been growing with time, while at the same time radioactive decay is
decreasing, the two magically cancel out and throughout Earth's history, the crust has always produced 7 TW of power.

Added a SurfEnFluxThermint function, which is HflowUMan - RadPowerTot. I thought that the second term should be (RadPowerMan+
RadPowerCore), but that gives too large a value, whereas RadPowerTot is dead on. I'll leave that for now, but I'm worried about it.

I deprecated dTidePower from EQTIDE and the body struct member dTidalPowMan now contains that info. This is really only useful for
EQTIDE+THERMINT runs in which the tidal power is necessary for the geophysical evolution. Otherwise, it's just a derived quantity in EQTIDE.