The DistOrb+DistRot example in the testing directory is now mem-check clean! No unitinitialized parameters, no memory leaks.
I set dMeanA to 0 for all bodies in VerifyDistOrb prior to the call to inv_plane. I think thisis OK.

Many mallocked parameters were not freed. Always free your pointers at the end of the subroutine! When you malloc it, immediately go to the end of the subroutine and free it, then fill in the code in between.