All testing examples are now memcheck clean!! The only issue is that valgrind is complaining that the mallocs in InitializeControlEqtide are a memory leak. I'm not sure why this is, because those mallocs need to happen, and other places where members of structs are mallocked are not causing issues.
In Binary, I had to initialize dR0 for the primary (0) and secondary (dSemi). I think that those choices are OK.

HZLimitDryRunaway is an interesting output parameter as it depends on other bodies. At this time, I just do a check on if it is a planet or a star. If the former, then the calculation is made. If the latter, the output is -1.