Moved Proxima b files around to prepare for resubmission and code release. In the tex/ directory, the Submit subdirectory contains the original manuscript as submitted to Astrobiology. The new Revised subdirectory contains our revisions.
The tex/Revised/Figures subdirectory contains an example of how I'd like to see the figures produces. This is the HZ evolution figure, and it contains the input files, the output files, and the python script that produces the PDF. Note that all these files should be documented and with names that are sensisble. I also include a README.