Working to separate thermint and eqtide...
Removed all mentions of tides from thermint

Moved fdHflowSecMan to body.c. This function now collates power from both thermint and eqtide in a new form of module coupling. RadHeat should also be separated like this, but that's not done yet.

Created new tide model "DB15" for Driscoll & Barnes (2015). The tidal subroutines that Peter put into thermint will constitute the core of this new model.

I also moved the tidal elements in the body struct from the section on thermint to that of eqtide.