Massive structural changes!
New fn_ptr VerifyImK2Module, part of control. Now every module can check that the options related to tides, through Im(k2) have been entered correctly.

New funtion VerifyImK2 has been added to VerifyOptions. This new function sprays out fn calls to all the modules to verify Im(k2) inputs are correct, following the procedure for PropsAux. I expect this will be a paradigm for future cross-module properties.

Revision to the main ReadOptions algorithm: Now the code initializes control->fnVerifyProperty (ImK2) as part of the ReadInitialOptions function. Then the assignments of control->fnVerifyProperty can be included in the AddModule calls in FinalizeModule.

This version is copmiling, but is not completed.