Updated functions in magmoc to match protoypes.
Code compiles, all tests pass except WaterLossCPL, and is probably related to the lines in module.c that were in conflict. I think the solution is to simpy remove line module.c:1022. On the other hand, we may need to think more carefully about how we want to calculate tidal heating in a magma ocean.